backend/ml_inference/completion_verifier.py
from PIL import Image
import numpy as np

class CompletionVerifier:
    def __init__(self, model_path='models/siamese_model.h5'):
        try:
            # self.model = tf.keras.models.load_model(model_path)
            self.model = None
        except:
            self.model = None
    
    def verify_completion(self, before_image_path, after_image_path):
        """Verify if issue is actually resolved by comparing before/after images"""
        if self.model is None:
            return self.manual_verification_required()
        
        # TODO: Implement Siamese network-based comparison
        # Load and preprocess images
        # Compare using trained model
        return {
            'verified': True,
            'confidence': 0.85,
            'requires_manual_review': False
        }
    
    def manual_verification_required(self):
        """Return flag indicating manual verification is needed"""
        return {
            'verified': False,
            'confidence': 0.0,
            'requires_manual_review': True
        }
    
    def calculate_similarity(self, img1, img2):
        """Calculate similarity score between two images"""
        # TODO: Implement image similarity calculation
        pass
    
    def detect_changes(self, before_image, after_image):
        """Detect changes between before and after images"""
        # TODO: Implement change detection algorithm
        pass
