backend/requirements.txt
Flask==2.3.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.2
python-dotenv==1.0.0
psycopg2-binary==2.9.7
pymongo==4.5.0
redis==5.0.0
celery==5.3.1
Pillow==10.0.0
requests==2.31.0
numpy==1.24.3
tensorflow==2.13.0
scikit-learn==1.3.0
geopy==2.3.0
boto3==1.28.25
firebase-admin==6.2.0
